Refactoring edc-tests path
