+++ edc-tests/README.md
@@ -1,7 +1,7 @@
 # Invoke Business-Tests via Maven
 
 ```shell
-./mvnw -pl edc-tests -Pbusiness-tests -pl edc-tests test -Dtest=net.catenax.edc.tests.features.RunCucumberTest
+./mvnw -pl edc-tests -Pbusiness-tests -pl edc-tests test -Dtest=RunCucumberTest
 ```
 
 ## Test locally using Act Tool
