--- edc-tests/src/test/java/net/catenax/edc/tests/BackendServiceSteps.java
@@ -1,14 +1,0 @@
-package net.catenax.edc.tests;
-
-import io.cucumber.java.en.Given;
-
-public class BackendServiceSteps {
-
-  @Given("'{connector}' has an empty backend-service")
-  public void cleanBackendService(Connector connector) {
-    final BackendServiceBackendAPI backendServiceBackendAPI =
-        connector.getBackendServiceBackendAPI();
-
-    backendServiceBackendAPI.list("/").forEach(backendServiceBackendAPI::delete);
-  }
-}
