--- edc-tests/src/test/java/net/catenax/edc/tests/data/BusinessPartnerNumberConstraint.java
@@ -1,10 +1,0 @@
-package net.catenax.edc.tests.data;
-
-import lombok.NonNull;
-import lombok.Value;
-
-@Value
-public class BusinessPartnerNumberConstraint implements Constraint {
-
-  @NonNull String businessPartnerNumber;
-}
