--- edc-tests/src/test/java/net/catenax/edc/tests/data/Constraint.java
@@ -1,3 +1,0 @@
-package net.catenax.edc.tests.data;
-
-public interface Constraint {}
