--- edc-tests/src/test/java/net/catenax/edc/tests/data/ContractNegotiationState.java
@@ -1,23 +1,0 @@
-/*
- *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Mercedes-Benz Tech Innovation GmbH - Initial Implementation
- *
- */
-
-package net.catenax.edc.tests.data;
-
-public enum ContractNegotiationState {
-  UNKNOWN,
-  INITIAL,
-  DECLINED,
-  CONFIRMED,
-  ERROR
-}
