+++ edc-tests/src/test/java/org/eclipse/tractusx/edc/tests/BackendServiceSteps.java
@@ -1,0 +1,14 @@
+package org.eclipse.tractusx.edc.tests;
+
+import io.cucumber.java.en.Given;
+
+public class BackendServiceSteps {
+
+  @Given("'{connector}' has an empty backend-service")
+  public void cleanBackendService(Connector connector) {
+    final BackendServiceBackendAPI backendServiceBackendAPI =
+        connector.getBackendServiceBackendAPI();
+
+    backendServiceBackendAPI.list("/").forEach(backendServiceBackendAPI::delete);
+  }
+}
