+++ edc-tests/src/test/java/org/eclipse/tractusx/edc/tests/Constants.java
@@ -1,0 +1,29 @@
+/*
+ *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Mercedes-Benz Tech Innovation GmbH - Initial API and Implementation
+ *
+ */
+
+package org.eclipse.tractusx.edc.tests;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public final class Constants {
+  public static final String DATA_MANAGEMENT_URL = "DATA_MANAGEMENT_URL";
+  public static final String DATA_MANAGEMENT_API_AUTH_KEY = "DATA_MANAGEMENT_API_AUTH_KEY";
+  public static final String IDS_URL = "IDS_URL";
+  public static final String DATA_PLANE_URL = "DATA_PLANE_URL";
+  public static final String BACKEND_SERVICE_BACKEND_API_URL = "BACKEND_SERVICE_BACKEND_API_URL";
+  public static final String DATABASE_URL = "DATABASE_URL";
+  public static final String DATABASE_USER = "DATABASE_USER";
+  public static final String DATABASE_PASSWORD = "DATABASE_PASSWORD";
+}
