+++ edc-tests/src/test/java/org/eclipse/tractusx/edc/tests/data/BusinessPartnerNumberConstraint.java
@@ -1,0 +1,10 @@
+package org.eclipse.tractusx.edc.tests.data;
+
+import lombok.NonNull;
+import lombok.Value;
+
+@Value
+public class BusinessPartnerNumberConstraint implements Constraint {
+
+  @NonNull String businessPartnerNumber;
+}
