+++ edc-tests/src/test/java/org/eclipse/tractusx/edc/tests/data/Constraint.java
@@ -1,0 +1,3 @@
+package org.eclipse.tractusx.edc.tests.data;
+
+public interface Constraint {}
