+++ edc-tests/src/test/java/org/eclipse/tractusx/edc/tests/data/ContractDefinition.java
@@ -1,0 +1,29 @@
+/*
+ *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Mercedes-Benz Tech Innovation GmbH - Initial API and Implementation
+ *
+ */
+package org.eclipse.tractusx.edc.tests.data;
+
+import java.util.List;
+import lombok.NonNull;
+import lombok.Value;
+
+@Value
+public class ContractDefinition {
+
+  @NonNull String id;
+
+  @NonNull String contractPolicyId;
+  @NonNull String acccessPolicyId;
+
+  List<String> assetIds;
+}
