+++ edc-tests/src/test/resources/org/eclipse/tractusx/edc/tests/features/ContractNegotiation.feature
